Added small exercise with animated smiley faces
There are two versions to this exercise.  The first is how the school coded it.  It the second (enhanced) version, I changed to colors to my liking, made the smiley face open its eyes and expand slightly on mouse hover.  Then on mouse click and hold, it bounces up and down.  Each face demonstrates different bezier function parameters.